refactor: editorDialog imgUrl validate rule
