+++ src/app/(global)/(dashboard)/article/publish/page.tsx
@@ -18,6 +18,8 @@ import { createArticle } from '@/lib/authApi';
 import { toast } from '@/components/ui/use-toast';
 import { Tag, TagInput } from 'emblor';
 import { getCategories } from '@/lib/nextApi';
+import Link from 'next/link';
+import { urlRegex } from '@/constants/imgUrlValidate';
 
 const isNeedPayOptions: { id: string; name: string; }[]= [
   { id: '免費', name: '免費' },
@@ -42,7 +44,7 @@ export default function PublishArticle() {
   const formSchema = z.object({
     title: z.string().min(1, { message: '標題是必填欄位' }).max(30, { message: '標題不能超過60個字' }),
     abstract: z.string().max(150, { message: '摘要不能超過150個字' }),
-    thumbnailUrl: z.string().optional().refine(val => !val || val.startsWith('https://'), { message: 'URL must start with https' }),
+    thumbnailUrl: z.string().optional().refine(val => !val || urlRegex.test(val), { message: '請至 imgur 或 unsplash 上傳圖片，造成不便敬請見諒' }),
     category: categoryValidation,
     isNeedPay: isNeedPayValidation,
     // category: z.string().refine(value => categoryOptions.some(option => option.id === value), {
@@ -79,7 +81,21 @@ export default function PublishArticle() {
     if (!editorProps) return;
     const tags = Array.isArray(values.tags) ? values.tags.map(tag => tag.text) : [];
     const isNeedPay = (values.isNeedPay === '免費') ? false : true;
-    const createArticleRequest = { ...editorProps, ...values, tags, isNeedPay };
+
+    let createArticleRequest;
+
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    const { thumbnailUrl, ...restEditorProps } = { ...editorProps };
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    const { thumbnailUrl: _, ...restValues } = { ...values };
+
+    if (values.thumbnailUrl && values.thumbnailUrl.trim()) {
+      createArticleRequest = { ...restEditorProps, ...values, tags, isNeedPay };
+    } else {
+      createArticleRequest = { ...restEditorProps, ...restValues, tags, isNeedPay };
+    }
+    console.log(createArticleRequest);
+
     createArticleMutate(createArticleRequest, {
       onSuccess: () => {
         toast({ title: '成功送出', variant: 'success' });
@@ -183,7 +199,7 @@ export default function PublishArticle() {
               )}
             />
             <div className='text-center'>
-              <Button className='mr-4 bg-grey text-black hover:bg-grey-400 hover:text-white'>取消</Button>
+              <Button className='mr-4 bg-grey text-black hover:bg-grey-300 hover:text-white' asChild><Link href='/manage/content'>取消</Link></Button>
               <Button type='submit' disabled={!isValid || !editorProps} isLoading={isUpdateCreateArticle}>發布文章</Button>
             </div>
           </form>
